refactor(rust): split model serde tests
Move marketplace parser coverage into a dedicated test file and share the fixture assertion helpers through test support to keep the serde test module shorter.
